+++ sdk/storage/azure-storage-file-datalake/src/samples/java/com/azure/storage/file/datalake/FileSystemAsyncClientJavaDocCodeSamples.java
@@ -222,7 +222,7 @@ public class FileSystemAsyncClientJavaDocCodeSamples {
         String owner = "rwx";
         String group = "r--";
         String leaseId = UUID.randomUUID().toString();
-        Long duration = 15L;
+        Integer duration = 15;
         DataLakePathCreateOptions options = new DataLakePathCreateOptions()
             .setPermissions(permissions)
             .setUmask(umask)
@@ -302,7 +302,7 @@ public class FileSystemAsyncClientJavaDocCodeSamples {
         String owner = "rwx";
         String group = "r--";
         String leaseId = UUID.randomUUID().toString();
-        Long duration = 15L;
+        Integer duration = 15;
 
         DataLakePathCreateOptions options = new DataLakePathCreateOptions()
             .setPermissions(permissions)
